refactor(NgZone): Remove elegation of NgZone to NgZoneImpl.
Align NgZone API a bit with the internal ManagedZone interface.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=140641923
